# doc/apidocs.swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This documentation describes all the MC APIs available",
    "title": "Master Controller (MC) API Documentation",
    "version": "0.0.1"
  },
  "basePath": "/api/v1",
  "paths": {
    "/auth/user/delete": {
      "post": {
        "description": "Delete a user",
        "tags": [
          "User"
        ],
        "operationId": "DeleteUser",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/success"
          },
          "400": {
            "$ref": "#/responses/badRequest"
          },
          "403": {
            "$ref": "#/responses/forbidden"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        }
      }
    },
    "/auth/user/show": {
      "post": {
        "description": "Show all users",
        "tags": [
          "User"
        ],
        "operationId": "ShowUser",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Organization"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/success"
          },
          "400": {
            "$ref": "#/responses/badRequest"
          },
          "403": {
            "$ref": "#/responses/forbidden"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "description": "Login to MC",
        "tags": [
          "Setup"
        ],
        "operationId": "Login",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/success"
          },
          "400": {
            "$ref": "#/responses/badRequest"
          },
          "403": {
            "$ref": "#/responses/forbidden"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        }
      }
    },
    "/usercreate": {
      "post": {
        "description": "Create new user",
        "tags": [
          "User"
        ],
        "operationId": "CreateUser",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CreateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/success"
          },
          "400": {
            "$ref": "#/responses/badRequest"
          },
          "403": {
            "$ref": "#/responses/forbidden"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        }
      }
    }
  },
  "definitions": {
...
    "CreateUser": {
      "type": "object",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "Email": {
          "type": "string"
        },
        "EmailVerified": {
          "type": "boolean"
        },
        "FamilyName": {
          "type": "string"
        },
        "GivenName": {
          "type": "string"
        },
        "Iter": {
          "type": "integer",
          "format": "int64"
        },
        "Locked": {
          "type": "boolean"
        },
        "Name": {
          "type": "string"
        },
        "Nickname": {
          "type": "string"
        },
        "Passhash": {
          "type": "string"
        },
        "Picture": {
          "type": "string"
        },
        "Salt": {
          "type": "string"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "verify": {
          "$ref": "#/definitions/EmailRequest"
        }
      },
      "x-go-package": "github.com/mobiledgex/edge-cloud-infra/mc/ormapi"
    },
...
    "EmailRequest": {
      "description": "Email request is used for password reset and to resend welcome\nverification email. It contains the information need to send\nsome kind of email to the user.",
      "type": "object",
      "properties": {
        "browser": {
          "type": "string",
          "x-go-name": "Browser"
        },
        "callbackurl": {
          "type": "string",
          "x-go-name": "CallbackURL"
        },
        "clientip": {
          "type": "string",
          "x-go-name": "ClientIP"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "operatingsystem": {
          "type": "string",
          "x-go-name": "OperatingSystem"
        }
      },
      "x-go-package": "github.com/mobiledgex/edge-cloud-infra/mc/ormapi"
    },
...
    "Organization": {
      "type": "object",
      "properties": {
        "Address": {
          "type": "string"
        },
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "Name": {
          "type": "string"
        },
        "Phone": {
          "type": "string"
        },
        "PublicImages": {
          "type": "boolean"
        },
        "Type": {
          "type": "string"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "github.com/mobiledgex/edge-cloud-infra/mc/ormapi"
    },
...
    "Result": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/mobiledgex/edge-cloud-infra/mc/ormapi"
    },
...
    "User": {
      "type": "object",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "Email": {
          "type": "string"
        },
        "EmailVerified": {
          "type": "boolean"
        },
        "FamilyName": {
          "type": "string"
        },
        "GivenName": {
          "type": "string"
        },
        "Iter": {
          "type": "integer",
          "format": "int64"
        },
        "Locked": {
          "type": "boolean"
        },
        "Name": {
          "type": "string"
        },
        "Nickname": {
          "type": "string"
        },
        "Passhash": {
          "type": "string"
        },
        "Picture": {
          "type": "string"
        },
        "Salt": {
          "type": "string"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "github.com/mobiledgex/edge-cloud-infra/mc/ormapi"
    },
    "UserLogin": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "github.com/mobiledgex/edge-cloud-infra/mc/ormapi"
    }
  },
  "responses": {
    "badRequest": {
      "description": "Status Bad Request",
      "schema": {
        "$ref": "#/definitions/Result"
      }
    },
    "forbidden": {
      "description": "Forbidden",
      "schema": {
        "$ref": "#/definitions/Result"
      }
    },
    "notFound": {
      "description": "Not Found",
      "schema": {
        "$ref": "#/definitions/Result"
      }
    },
    "success": {
      "description": "Success",
      "schema": {
        "$ref": "#/definitions/Result"
      }
    }
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}